do minor change in the code. Change the float to double to match the numpy precision of the norm distance

+++ graph_pkg/edit_cost/edit_cost.pyx
@@ -9,7 +9,7 @@ cdef class EditCost:
     cpdef float cost_delete_node(self, Node node) except? -1:
         raise NotImplementedError
 
-    cpdef float cost_substitute_node(self, Node node1, Node node2) except? -1:
+    cpdef double cost_substitute_node(self, Node node1, Node node2) except? -1:
         raise NotImplementedError
 
     cpdef float cost_insert_edge(self, Edge edge) except? -1:
