+++ graph_pkg/graph/label/label_node_letter.pxd
@@ -2,7 +2,7 @@ from graph_pkg.graph.label.label_base cimport LabelBase
 
 cdef class LabelNodeLetter(LabelBase):
     cdef:
-        float x
-        float y
+        double x
+        double y
 
     cpdef tuple get_attributes(self)
